requirements.txt
Flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
pytest==7.4.3
numpy==1.26.4
scipy==1.11.4
statsmodels==0.14.1
gunicorn==21.2.0
